Fix wave file 96K sample rate for MH3-4531.

+++ apps/audio_enc/firmware/src/config/audio_enc_sam_e70_xult_wm8904_i2sc_freertos/decoder/audio_decoders/wav/wav_dec.c
@@ -95,7 +95,7 @@ unsigned int WAV_HdrGetFileSize(void)
     return (unsigned int) wavHeader.extralen;
 }
 
-uint16_t WAV_GetSampleRate(void){
+uint32_t WAV_GetSampleRate(void){
     return wavDecoder.wavSampleRate;
 }
